core: align network stats tx\rx_bytes data types
Data type mismatches of network statistics between vdsm, engine bll and
postgres db caused overflow exceptions in engine and the db:
- vdsm: unsigned 64 (U64), max value: 9E+19 [1][2]
- engine: Long, max value: 9E+18
- db: DECIMAL(18,4), max value: 1.2345E14

This patch changes engine and db data types to:
- engine: BigInteger, max value: 2^Integer.MAX_VALUE
- db: NUMERIC(20, 0), max value: 1E+20 for whole numbers
      NUMERIC(24, 4), max value: 9.1234E19 for others

[1] https://www.kernel.org/doc/html/latest/networking/
    statistics.html#c.rtnl_link_stats64
[2] https://github.com/oVirt/vdsm/blob/
    51ee47b235ec28744165190a22e211d43fb371b7/lib/vdsm/api/
    vdsm-api.yml#L5076

Bug-Url: https://bugzilla.redhat.com/1980192

+++ backend/manager/modules/common/src/main/java/org/ovirt/engine/core/common/businessentities/network/NetworkStatistics.java
@@ -1,5 +1,6 @@
 package org.ovirt.engine.core.common.businessentities.network;
 
+import java.math.BigInteger;
 import java.util.Objects;
 
 import org.ovirt.engine.core.common.businessentities.BusinessEntityWithStatus;
@@ -20,17 +21,17 @@ public abstract class NetworkStatistics implements BusinessEntityWithStatus<Guid
 
     private Double receiveRate;
 
-    private Long receivedBytes;
+    private BigInteger receivedBytes;
 
-    private Long receivedBytesOffset;
+    private BigInteger receivedBytesOffset;
 
     private Double transmitDropRate;
 
     private Double transmitRate;
 
-    private Long transmittedBytes;
+    private BigInteger transmittedBytes;
 
-    private Long transmittedBytesOffset;
+    private BigInteger transmittedBytesOffset;
 
     private Double sampleTime;
 
@@ -135,7 +136,7 @@ public abstract class NetworkStatistics implements BusinessEntityWithStatus<Guid
      * @param receivedBytes
      *            the total received bytes.
      */
-    public void setReceivedBytes(Long receivedBytes) {
+    public void setReceivedBytes(BigInteger receivedBytes) {
         this.receivedBytes = receivedBytes;
     }
 
@@ -144,7 +145,7 @@ public abstract class NetworkStatistics implements BusinessEntityWithStatus<Guid
      *
      * @return the total received bytes.
      */
-    public Long getReceivedBytes() {
+    public BigInteger getReceivedBytes() {
         return receivedBytes;
     }
 
@@ -154,7 +155,7 @@ public abstract class NetworkStatistics implements BusinessEntityWithStatus<Guid
      * @param receivedBytesOffset
      *            the RX byte offset.
      */
-    public void setReceivedBytesOffset(Long receivedBytesOffset) {
+    public void setReceivedBytesOffset(BigInteger receivedBytesOffset) {
         this.receivedBytesOffset = receivedBytesOffset;
     }
 
@@ -163,7 +164,7 @@ public abstract class NetworkStatistics implements BusinessEntityWithStatus<Guid
      *
      * @return the RX byte offset.
      */
-    public Long getReceivedBytesOffset() {
+    public BigInteger getReceivedBytesOffset() {
         return receivedBytesOffset;
     }
 
@@ -173,7 +174,7 @@ public abstract class NetworkStatistics implements BusinessEntityWithStatus<Guid
      * @param transmittedBytesOffset
      *            the TX byte offset.
      */
-    public void setTransmittedBytesOffset(Long transmittedBytesOffset) {
+    public void setTransmittedBytesOffset(BigInteger transmittedBytesOffset) {
         this.transmittedBytesOffset = transmittedBytesOffset;
     }
 
@@ -182,7 +183,7 @@ public abstract class NetworkStatistics implements BusinessEntityWithStatus<Guid
      *
      * @return the TX byte offset.
      */
-    public Long getTransmittedBytesOffset() {
+    public BigInteger getTransmittedBytesOffset() {
         return transmittedBytesOffset;
     }
 
@@ -230,7 +231,7 @@ public abstract class NetworkStatistics implements BusinessEntityWithStatus<Guid
      * @param transmittedBytes
      *            the total transmitted bytes.
      */
-    public void setTransmittedBytes(Long transmittedBytes) {
+    public void setTransmittedBytes(BigInteger transmittedBytes) {
         this.transmittedBytes = transmittedBytes;
     }
 
@@ -239,7 +240,7 @@ public abstract class NetworkStatistics implements BusinessEntityWithStatus<Guid
      *
      * @return the total transmitted bytes.
      */
-    public Long getTransmittedBytes() {
+    public BigInteger getTransmittedBytes() {
         return transmittedBytes;
     }
