+++ frontend/webadmin/modules/uicommonweb/src/main/java/org/ovirt/engine/ui/uicommonweb/models/hosts/HostInterface.java
@@ -1,5 +1,7 @@
 package org.ovirt.engine.ui.uicommonweb.models.hosts;
 
+import java.math.BigInteger;
+
 import org.ovirt.engine.core.common.businessentities.network.InterfaceStatus;
 import org.ovirt.engine.core.common.businessentities.network.VdsNetworkInterface;
 import org.ovirt.engine.ui.uicommonweb.models.Model;
@@ -106,23 +108,23 @@ public class HostInterface extends Model {
         txRate = value;
     }
 
-    private Long rxTotal;
+    private BigInteger rxTotal;
 
-    public Long getRxTotal() {
+    public BigInteger getRxTotal() {
         return rxTotal;
     }
 
-    public void setRxTotal(Long value) {
+    public void setRxTotal(BigInteger value) {
         rxTotal = value;
     }
 
-    private Long txTotal;
+    private BigInteger txTotal;
 
-    public Long getTxTotal() {
+    public BigInteger getTxTotal() {
         return txTotal;
     }
 
-    public void setTxTotal(Long value) {
+    public void setTxTotal(BigInteger value) {
         txTotal = value;
     }
